fix: adding redirect to 500 and updating unit tests
