+++ apps/web/src/__tests__/RequestSupport.spec.tsx
@@ -1,37 +1,43 @@
-import { NextSeo } from 'next-seo'
+import type { Document } from '@contentful/rich-text-types'
+import { BLOCKS } from '@contentful/rich-text-types'
+import { screen } from '@testing-library/react'
+import type { GetServerSidePropsContext } from 'next'
 
-import { render, screen } from '@/config/TestUtils'
+import { render } from '@/config/TestUtils'
+import RequestSupport, { getServerSideProps } from '@/pages/request-support/index'
+import { RichTextRenderer } from '@/utils/Renderers/RichTextRenderer/RichTextRenderer'
 
-import RequestSupport from '../pages/request-support'
+jest.mock('@nihr-ui/frontend', () => ({
+  Container: jest.fn(({ children }) => <div>{children}</div>),
+}))
 
-jest.mock('next-seo')
+jest.mock('@/utils/Renderers/RichTextRenderer/RichTextRenderer', () => ({
+  RichTextRenderer: jest.fn(({ children }) => <div>{JSON.stringify(children)}</div>),
+}))
 
-describe('RequestSupport Component', () => {
-  it('renders page title and heading correctly', () => {
-    render(<RequestSupport />)
-    expect(NextSeo).toHaveBeenCalledWith({ title: `Request NIHR RDN support` }, {})
-    expect(screen.getByRole('heading', { name: /Request NIHR RDN support/i })).toBeInTheDocument()
-  })
+jest.mock('@/lib/contentful/contentfulService', () => ({
+  getPage: jest.fn(),
+  getBporFooter: jest.fn(),
+}))
 
-  it('renders link to local RDN', () => {
-    render(<RequestSupport />)
-    const link = screen.getByRole('link', { name: /your local network/i })
-    expect(link).toHaveAttribute('href', 'https://www.nihr.ac.uk/documents/study-support-service-contacts/11921')
-  })
+jest.mock('next/router', () => ({
+  useRouter: jest.fn(),
+}))
 
-  it('renders list items correctly', () => {
-    render(<RequestSupport />)
+describe('RequestSupportPage', () => {
+  it('renders RichTextRenderer for content and footerContent', () => {
+    const richTextContent: Document = {
+      nodeType: BLOCKS.DOCUMENT,
+      content: [],
+      data: {},
+    }
 
-    const listItems = [
-      'Supporting study-wide planning activities (e.g. identification of potential sites), support with cost attribution and the identification of resource requirements using UK wide tools (SoECAT and iCT)',
-      'Research delivery advice (e.g. recruitment strategies, regional care pathways and NHS support services, investigators, capabilities, etc.)',
-      'Discuss site issues, including study delivery',
-      'Advice regarding engagement with local communities',
-      'Clinical advice (e.g. assessment of study deliverability in the UK, recommendations of recruitment methods and pathways, study design considerations)',
-      'Support to overcome barriers, challenges or other types of requests',
-    ]
-
-    expect(screen.getAllByRole('listitem').map((listItem) => listItem.textContent)).toEqual(listItems)
+    const { getAllByText } = render(
+      <RequestSupport content={{ richText: richTextContent }} footerContent={{ richText: richTextContent }} />
+    )
+    const elements = getAllByText(JSON.stringify(richTextContent))
+    expect(elements).toHaveLength(2)
+    expect(RichTextRenderer).toHaveBeenCalledWith(expect.objectContaining({ children: richTextContent }), {})
   })
 
   it('renders link to previous page', () => {
@@ -43,5 +49,21 @@ describe('RequestSupport Component', () => {
   it('does not render link to previous page if no referer', () => {
     render(<RequestSupport />)
     expect(screen.queryByRole('link', { name: /Return to previous page/i })).not.toBeInTheDocument()
+  })
+})
+
+describe('getServerSideProps', () => {
+  it('redirects to /500 if content and footerContent are missing', async () => {
+    const context = {
+      query: {},
+    } as unknown as GetServerSidePropsContext
+
+    const result = await getServerSideProps(context)
+
+    expect(result).toEqual({
+      redirect: {
+        destination: '/500',
+      },
+    })
   })
 })
